+++ resources/[dg]/dg-admin/typescript/src/server/modules/commands/data/spawnVehicle.ts
@@ -1,5 +1,3 @@
-import { Notifications } from '@dgx/server';
-
 import { Inputs } from '../../../enums/inputs';
 
 declare interface SpawnVehicleData {